Partial pass of init file breakout
