+++ slackest/__init__.py
@@ -16,17 +16,10 @@ import datetime
 import json
 import time
 
-import requests
-
 from slackest.utils import get_item_id_by_name
 
 
-API_BASE_URL = 'https://slack.com/api/{api}'
-DEFAULT_TIMEOUT = 10
-DEFAULT_RETRIES = 0
-# seconds to wait after a 429 error if Slack's API doesn't provide one
-DEFAULT_WAIT = 20
-DEFAULT_API_SLEEP = 5
+from .constants import *
 
 __version__ = '0.13.3'
 __all__ = ['SlackestError', 'Response', 'BaseAPI', 'API', 'Auth', 'Users',
@@ -38,2137 +31,51 @@ __all__ = ['SlackestError', 'Response', 'BaseAPI', 'API', 'Auth', 'Users',
            'AppsPermissions', 'Slackest', 'Dialog']
 
 
-class SlackestError(Exception):
-    """Dummy exception placeholder"""
-    pass
+from .slackest_error import SlackestError
 
+from .response import Response
 
-class Response(object):
-    """Requests response object"""
+from .base_api import BaseAPI
 
-    def __init__(self, body):
-        self.raw = body
-        self.body = json.loads(body)
-        self.successful = self.body['ok']
-        self.error = self.body.get('error')
+from .api import API
 
-    def __str__(self):
-        return json.dumps(self.body)
+from .auth import Auth
 
+from .dialog import Dialog
 
-class BaseAPI(object):
-    """BaseAPI interface for making the `requests` calls to Slack."""
+from .users_profile import UsersProfile
 
-    def __init__(self, token=None, timeout=DEFAULT_TIMEOUT, proxies=None,
-                 session=None, rate_limit_retries=DEFAULT_RETRIES):
-        self.token = token
-        self.timeout = timeout
-        self.proxies = proxies
-        self.session = session
-        self.rate_limit_retries = rate_limit_retries
-        self.retry_max = 10
-        self.retry_counter = 0
-        self.retry_wait_secs = 1
+from .users_admin import UsersAdmin
 
-    def _request(self, method, api, **kwargs):
-        """
-        Internal request call, with rate limiting retries
+from .users import Users
 
-        :param method: The method to call: GET, POST, etc.
-        :type method: :class:`Request <Request>` object
-        :param api: The API endpoint
-        :type api: str
-        :param kwargs: Various keyword arguments that could be passed to the request
-        :type kwargs: dict
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        if self.token:
-            kwargs.setdefault('params', {})['token'] = self.token
+from .groups import Groups
 
-        # while we have rate limit retries left, fetch the resource and back
-        # off as Slack's HTTP response suggests
-        for retry_num in range(self.rate_limit_retries):
-            response = method(API_BASE_URL.format(api=api),
-                              timeout=self.timeout,
-                              proxies=self.proxies,
-                              **kwargs)
+from .channels import Channels
 
-            if response.status_code == requests.codes.ok:
-                break
+from .conversation import Conversation
 
-            # handle HTTP 429 as documented at
-            # https://api.slack.com/docs/rate-limits
-            elif response.status_code == requests.codes.too_many: # HTTP 429
-                time.sleep(int(response.headers.get('retry-after', DEFAULT_WAIT)))
-                continue
-            else:
-                response.raise_for_status()
-        else:
-            # with no retries left, make one final attempt to fetch the resource,
-            # but do not handle too_many status differently
-            response = method(API_BASE_URL.format(api=api),
-                              timeout=self.timeout,
-                              proxies=self.proxies,
-                              **kwargs)
-            response.raise_for_status()
+from .chat import Chat
 
-        response = Response(response.text)
-        if not response.successful:
-            raise SlackestError(response.error)
+from .im import IM
 
-        return response
+from .mpim import MPIM
 
-    def _session_get(self, url, params=None, **kwargs):
-        """
-        Internal request GET call with session
+from .search import Search
 
-        :param url: The URL to request
-        :type url: str
-        :param params: Dictionary containing URL request parameters (headers, etc.)
-        :type params: dict
-        :param kwargs: Various keyword arguments that could be passed to the GET
-        :type kwargs: dict
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        kwargs.setdefault('allow_redirects', True)
-        return self.session.request(
-            method='get', url=url, params=params, **kwargs
-        )
+from .files_comments import FilesComments
 
-    def _session_post(self, url, data=None, **kwargs):
-        """
-        Internal request POST call with session
+from .files import Files
 
-        :param url: The URL to request
-        :type url: str
-        :param params: Dictionary containing request data
-        :type params: dict
-        :param kwargs: Various keyword arguments that could be passed to the POST
-        :type kwargs: dict
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.session.request(
-            method='post', url=url, data=data, **kwargs
-        )
+from .stars import Stars
 
-    def get(self, api, **kwargs):
-        """
-        External request GET call, wraps around internal _request.get
+from .emoji import Emoji
 
-        :param api: The API endpoint to connect to
-        :type api: str
-        :param kwargs: Various keyword arguments that could be passed to the GET
-        :type kwargs: dict
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        try:
-            return self._request(
-                self._session_get if self.session else requests.get,
-                api, **kwargs
-            )
-        except requests.HTTPError as e:
-            # Put a retry here with a short circuit
-            if self.retry_counter < self.retry_max:
-                self.retry_counter = self.retry_counter + 1
-                time.sleep(self.retry_wait_secs)
-                self.get(api, **kwargs)
-            else:
-                raise
+from .presence import Presence
 
-    def post(self, api, **kwargs):
-        """
-        External request POST call, wraps around internal _request.post
+from .rtm import RTM
 
-        :param api: The API endpoint to connect to
-        :type api: str
-        :param kwargs: Various keyword arguments that could be passed to the POST
-        :type kwargs: dict
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self._request(
-            self._session_post if self.session else requests.post,
-            api, **kwargs
-        )
-
-
-class API(BaseAPI):
-    """Follows the Slack Test API. See https://api.slack.com/methods"""
-
-    def test(self, error=None, **kwargs):
-        """
-        Allows access to the Slack API test endpoint
-
-        :param error: The API error
-        :type error: str
-        :param kwargs: Various keyword arguments that could be passed to the request
-        :type kwargs: dict
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        if error:
-            kwargs['error'] = error
-
-        return self.get('api.test', params=kwargs)
-
-
-class Auth(BaseAPI):
-    """Follows the Slack Auth API. See https://api.slack.com/methods"""
-
-    def test(self):
-        """
-        Allows access to the Slack API test endpoint
-
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.get('auth.test')
-
-    def revoke(self, test=True):
-        """
-        Allows access to the Slack API test endpoint
-
-        :param test: Boolean to run the test
-        :type test: bool
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('auth.revoke', data={'test': int(test)})
-
-
-class Dialog(BaseAPI):
-    """Follows the Slack Dialog API. See https://api.slack.com/methods"""
-
-    def open(self, dialog, trigger_id):
-        """
-        Opens a dialog with a user
-
-        :param dialog: JSON-encoded dialog definition
-        :type dialog: json
-        :param trigger_id: The trigger to post to the user.
-        :type trigger_id: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('dialog.open',
-                         data={
-                             'dialog': json.dumps(dialog),
-                             'trigger_id': trigger_id,
-                         })
-
-
-class UsersProfile(BaseAPI):
-    """Follows the Slack UsersProfile API. See https://api.slack.com/methods"""
-
-    def get(self, user=None, include_labels=False):
-        """
-        Gets a Slack user's profile
-
-        :param user: User to retrieve profile info for
-        :type user: str
-        :param include_labels: Include labels for each ID in custom profile fields
-        :type include_labels: bool
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return super(UsersProfile, self).get(
-            'users.profile.get',
-            params={'user': user, 'include_labels': str(include_labels).lower()}
-        )
-
-    def set(self, user=None, profile=None, name=None, value=None):
-        """
-        Gets a Slack user's profile
-
-        :param user: ID of user to change
-        :type user: str
-        :param profile: Collection of key:value pairs presented as a URL-encoded JSON hash
-        :type profile: str
-        :param name: Name of a single key to set
-        :type name: str
-        :param value: Value to set a single key to
-        :type value: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('users.profile.set',
-                         data={
-                             'user': user,
-                             'profile': profile,
-                             'name': name,
-                             'value': value
-                         })
-
-
-class UsersAdmin(BaseAPI):
-    """Follows the Slack UsersAdmin API. See https://api.slack.com/methods"""
-
-    def invite(self, email, channels=None, first_name=None,
-               last_name=None, resend=True):
-        """
-        DEPRECATED - Invites a user to channel(s) via email. Looks to be deprecated.
-
-        :param email: Email of the user to invite to a channel(s)
-        :type email: str
-        :param channels: A CSV of channels for the invite.
-        :type channels: str
-        :param first_name: First name of the invitee
-        :type first_name: str
-        :param last_name: Last name of the invitee
-        :type last_name: str
-        :param resend: Whether or not this invite is a resend
-        :type resend: bool
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('users.admin.invite',
-                         params={
-                             'email': email,
-                             'channels': channels,
-                             'first_name': first_name,
-                             'last_name': last_name,
-                             'resend': resend
-                         })
-
-
-class Users(BaseAPI):
-    """Follows the Slack Users API. See https://api.slack.com/methods"""
-
-    def __init__(self, *args, **kwargs):
-        super(Users, self).__init__(*args, **kwargs)
-        self._profile = UsersProfile(*args, **kwargs)
-        self._admin = UsersAdmin(*args, **kwargs)
-
-    @property
-    def profile(self):
-        """
-        Returns the profile object attribute
-
-        :return: A usersprofile object.
-        :rtype: :class:`UsersProfile <UsersProfile>` object
-        """
-        return self._profile
-
-    @property
-    def admin(self):
-        return self._admin
-
-    def info(self, user, include_locale=False):
-        """
-        Returns information about the user
-
-        :param user: The Slack user ID of the user to look up
-        :type user: str
-        :param include_locale: Whether or not to include the user's locale
-        :type include_locale: bool
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.get('users.info',
-                        params={'user': user, 'include_locale': str(include_locale).lower()})
-
-    def list(self, cursor=None, include_locale=True, limit=500):
-        """
-        List all users in a Slack team.
-
-        :param cursor: Cursor pagination
-        :type cursor: str
-        :param include_locale: Receive the user's locale
-        :type include_locale: bool
-        :param limit: The maximum number of users to return
-        :type limit: int
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.get('users.list',
-                        params={'include_locale': str(include_locale).lower(),
-                                'limit': limit, 'cursor': cursor})
-
-    def list_all(self, include_locale=True):
-        """
-        Lists all users in a Slack team
-
-        :param include_locale: Receive the user's locale
-        :type include_locale: bool
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        response = self.get('users.list', params={'include_locale': str(include_locale).lower()})
-        members = response.body.get('members', [])
-        next_cursor = response.body.get('response_metadata', {}).get('next_cursor', '')
-        while next_cursor:
-            response = self.get('users.list',
-                                params={'include_locale': str(include_locale).lower(),
-                                        'cursor': next_cursor})
-            if response is not None:
-                members.extend(response.body.get('members', []))
-                next_cursor = response.body.get('response_metadata', {}).get('next_cursor', '')
-                time.sleep(DEFAULT_API_SLEEP)
-            else:
-                raise SlackestError("Null response received. You've probably hit the rate limit.")
-
-        if members:
-            response.body['members'] = members
-
-        return response
-
-    def identity(self):
-        """
-        Retrieves the user's identity: name, ID, team
-
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.get('users.identity')
-
-    def set_active(self):
-        """
-        Sets the user object to active. DEPRECATED
-
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('users.setActive')
-
-    def get_presence(self, user):
-        """
-        Gets the presence of the Slack user
-
-        :param user: The Slack user ID
-        :type user: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.get('users.getPresence', params={'user': user})
-
-    def set_presence(self, presence):
-        """
-        Sets the presence of the current Slack user
-
-        :param presence: The presence level of the user: either `auto` or `away`
-        :type presence: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('users.setPresence', data={'presence': presence})
-
-    def get_user_id(self, user_name):
-        """
-        Gets a user ID according to the user's name
-
-        :param user_name: The user's name
-        :type user_name: str
-        :return: Returns the user ID
-        :rtype: str
-        """
-        members = self.list_all().body['members']
-        return get_item_id_by_name(members, user_name)
-
-
-class Groups(BaseAPI):
-    """Follows the Slack Groups API. See https://api.slack.com/methods"""
-
-    def create(self, name):
-        """
-        Creates a group with the name
-
-        :param name: The group's name
-        :type name: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('groups.create', data={'name': name})
-
-    def create_child(self, channel):
-        """
-        Takes an existing private channel and performs the following steps:
-
-        * Renames the existing private channel (from "example" to "example-archived").
-        * Archives the existing private channel.
-        * Creates a new private channel with the name of the existing private channel.
-        * Adds all members of the existing private channel to the new private channel.
-
-        :param channel: Private channel to clone and archive
-        :type channel: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('groups.createChild', data={'channel': channel})
-
-    def info(self, channel):
-        """
-        Returns the private channel's information
-
-        :param channel: The private channel's ID
-        :type channel: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.get('groups.info', params={'channel': channel})
-
-    def list(self, exclude_archived=True, exclude_members=False):
-        """
-        Lists the private channels that the user has access to
-
-        :param exclude_archived: Don't include archived private channels in the returned list
-        :type exclude_archived: bool
-        :param exclude_members: Don't include members in the returned list
-        :type exclude_members: bool
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.get('groups.list',
-                        params={'exclude_archived': str(exclude_archived).lower(),
-                                'exclude_members': str(exclude_members).lower()})
-
-    def history(self, channel, latest=None, oldest=None, count=None,
-                inclusive=True):
-        """
-        Fetches history of messages and events from a private channel
-
-        :param channel: The private channel ID
-        :type channel: str
-        :param latest: End of time range to include in results
-        :type latest: str
-        :param oldest: Start of time range to include in results
-        :type oldest: str
-        :param count: The number of messages to return
-        :type count: int
-        :param inclusive: Include messages with latest or oldest timestamp in results
-        :type inclusive: bool
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.get('groups.history',
-                        params={
-                            'channel': channel,
-                            'latest': latest,
-                            'oldest': oldest,
-                            'count': count,
-                            'inclusive': int(inclusive)
-                        })
-
-    def invite(self, channel, user):
-        """
-        Invites a user to a private channel
-
-        :param channel: The private channel ID
-        :type channel: str
-        :param user: The user ID
-        :type user: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('groups.invite',
-                         data={'channel': channel, 'user': user})
-
-    def kick(self, channel, user):
-        """
-        Removes a user from a private channel
-
-        :param channel: The private channel ID
-        :type channel: str
-        :param user: The user ID
-        :type user: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('groups.kick',
-                         data={'channel': channel, 'user': user})
-
-    def leave(self, channel):
-        """
-        Allows a user object to remove themselves from a private channel
-
-        :param channel: The private channel ID
-        :type channel: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('groups.leave', data={'channel': channel})
-
-    def mark(self, channel, time_stamp):
-        """
-        Moves the read cursor in a private channel
-
-        :param channel: The private channel ID
-        :type channel: str
-        :param time_stamp: The timestamp of the most recently seen message
-        :type time_stamp: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('groups.mark', data={'channel': channel, 'ts': time_stamp})
-
-    def rename(self, channel, name):
-        """
-        Renames a private channel
-
-        :param channel: The private channel ID
-        :type channel: str
-        :param name: The new user-friendly name of the private channel
-        :type name: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('groups.rename',
-                         data={'channel': channel, 'name': name})
-
-    def replies(self, channel, thread_ts):
-        """
-        Retrieve a thread of messages posted to a private channel
-
-        :param channel: The private channel ID
-        :type channel: str
-        :param thread_ts: Unique identifier of a thread's parent message
-        :type thread_ts: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.get('groups.replies',
-                        params={'channel': channel, 'thread_ts': thread_ts})
-
-    def archive(self, channel):
-        """
-        Archives a private channel
-
-        :param channel: The private channel ID
-        :type channel: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('groups.archive', data={'channel': channel})
-
-    def unarchive(self, channel):
-        """
-        Unarchives a private channel
-
-        :param channel: The private channel ID
-        :type channel: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('groups.unarchive', data={'channel': channel})
-
-    def open(self, channel):
-        """
-        Opens a private channel
-
-        :param channel: The private channel ID
-        :type channel: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('groups.open', data={'channel': channel})
-
-    def close(self, channel):
-        """
-        Closes a private channel
-
-        :param channel: The private channel ID
-        :type channel: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('groups.close', data={'channel': channel})
-
-    def set_purpose(self, channel, purpose):
-        """
-        Sets the purpose of a private channel
-
-        :param channel: The private channel ID
-        :type channel: str
-        :param purpose: The purpose
-        :type purpose: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('groups.setPurpose',
-                         data={'channel': channel, 'purpose': purpose})
-
-    def set_topic(self, channel, topic):
-        """
-        Sets the topic of a private channel
-
-        :param channel: The private channel ID
-        :type channel: str
-        :param topic: The topic
-        :type topic: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('groups.setTopic',
-                         data={'channel': channel, 'topic': topic})
-
-
-class Channels(BaseAPI):
-    """Follows the Slack Channels API. See https://api.slack.com/methods"""
-
-    def create(self, name):
-        """
-        Creates a public channel
-
-        :param name: The name
-        :type name: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('channels.create', data={'name': name})
-
-    def info(self, channel):
-        """
-        Retrieves information about a public channel
-
-        :param name: The channel ID
-        :type name: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.get('channels.info', params={'channel': channel})
-
-    def list(self, exclude_archived=True, exclude_members=False):
-        """
-        Lists channels
-
-        :param exclude_archived: Exclude archived channels
-        :type exclude_archived: bool
-        :param exclude_members: Exclude members from being listed
-        :type exclude_members: bool
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.get('channels.list',
-                        params={'exclude_archived': str(exclude_archived).lower(),
-                                'exclude_members': str(exclude_members).lower()})
-
-    def history(self, channel, latest=None, oldest=None, count=None,
-                inclusive=False, unreads=False):
-        """
-        Fetches history of messages and events from a channel
-
-        :param channel: The channel ID
-        :type channel: str
-        :param latest: End of time range to include in results
-        :type latest: str
-        :param oldest: Start of time range to include in results
-        :type oldest: str
-        :param count: The number of messages to return
-        :type count: int
-        :param inclusive: Include messages with latest or oldest timestamp in results
-        :type inclusive: bool
-        :param unreads: Include `unread_count_display` in the output
-        :type unreads: bool
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.get('channels.history',
-                        params={
-                            'channel': channel,
-                            'latest': latest,
-                            'oldest': oldest,
-                            'count': count,
-                            'inclusive': int(inclusive),
-                            'unreads': int(unreads)
-                        })
-
-    def mark(self, channel, time_stamp):
-        """
-        Moves the read cursor in a public channel
-
-        :param channel: The channel ID
-        :type channel: str
-        :param time_stamp: The timestamp of the most recently seen message
-        :type time_stamp: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('channels.mark',
-                         data={'channel': channel, 'ts': time_stamp})
-
-    def join(self, name):
-        """
-        Allows a user object to join a channel
-
-        :param name: The channel name (#general)
-        :type name: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('channels.join', data={'name': name})
-
-    def leave(self, channel):
-        """
-        Allows a user object to leave a channel
-
-        :param name: The channel ID
-        :type name: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('channels.leave', data={'channel': channel})
-
-    def invite(self, channel, user):
-        """
-        Invites a user to a private channel
-
-        :param channel: The channel ID
-        :type channel: str
-        :param user: The user ID
-        :type user: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('channels.invite',
-                         data={'channel': channel, 'user': user})
-
-    def kick(self, channel, user):
-        """
-        Removes a user from a channel
-
-        :param channel: The private channel ID
-        :type channel: str
-        :param user: The user ID
-        :type user: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('channels.kick',
-                         data={'channel': channel, 'user': user})
-
-    def rename(self, channel, name):
-        """
-        Renames a channel
-
-        :param channel: The channel ID
-        :type channel: str
-        :param name: The new user-friendly name of the channel
-        :type name: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('channels.rename',
-                         data={'channel': channel, 'name': name})
-
-    def replies(self, channel, thread_ts):
-        """
-        Retrieve a thread of messages posted to a channel
-
-        :param channel: The channel ID
-        :type channel: str
-        :param thread_ts: Unique identifier of a thread's parent message
-        :type thread_ts: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.get('channels.replies',
-                        params={'channel': channel, 'thread_ts': thread_ts})
-
-    def archive(self, channel):
-        """
-        Archives a public channel
-
-        :param channel: The channel ID
-        :type channel: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('channels.archive', data={'channel': channel})
-
-    def unarchive(self, channel):
-        """
-        Unarchives a channel
-
-        :param channel: The channel ID
-        :type channel: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('channels.unarchive', data={'channel': channel})
-
-    def set_purpose(self, channel, purpose):
-        """
-        Sets the purpose of a channel
-
-        :param channel: The channel ID
-        :type channel: str
-        :param purpose: The purpose
-        :type purpose: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('channels.setPurpose',
-                         data={'channel': channel, 'purpose': purpose})
-
-    def set_topic(self, channel, topic):
-        """
-        Sets the topic of a channel
-
-        :param channel: The channel ID
-        :type channel: str
-        :param topic: The topic
-        :type topic: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('channels.setTopic',
-                         data={'channel': channel, 'topic': topic})
-
-    def get_channel_id(self, channel_name):
-        """
-        Gets a channel ID according to the channel's name
-
-        :param channel_name: The channel's name
-        :type channel_name: str
-        :return: Returns the channel ID
-        :rtype: str
-        """
-        channels = self.list().body['channels']
-        return get_item_id_by_name(channels, channel_name)
-
-
-class Conversation(BaseAPI):
-    """Follows the Slack Conversation API.
-    See https://api.slack.com/docs/conversations-api#methods"""
-
-    # A Python 2 and Python 3 compatible timestamp
-    now = datetime.datetime.now().timetuple()
-    timestamp = float((time.mktime(now)+datetime.datetime.now().microsecond/1000000.0))
-
-    def archive(self, channel):
-        """
-        Archives a channel
-
-        :param channel: The channel ID
-        :type channel: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('conversation.archive', data={'channel': channel})
-
-    def close(self, channel):
-        """
-        Closes a channel
-
-        :param channel: The channel ID
-        :type channel: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('conversations.close', data={'channel': channel})
-
-    def create(self, name, is_private=True, users=[]):
-        """
-        Creates a channel
-
-        :param name: The channel name
-        :type name: str
-        :param is_private: Determines if channel is private (like a group)
-        :type is_private: bool
-        :param users: A list of User IDs to add to the channel
-        :type users: list[str]
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        if isinstance(users, (tuple, list)):
-            users = ','.join(users)
-
-        return self.post('conversations.create',
-                         data={'name': name, 'is_private': str(is_private).lower(),
-                               'user_ids': users})
-
-    def history(self, channel, cursor=None, inclusive=False, limit=100,
-                latest=timestamp, oldest=0):
-        """
-        Fetches history of messages and events from a channel
-
-        :param channel: The channel ID
-        :type channel: str
-        :param cursor: the cursor id of the next set of history
-        :type cursor: str
-        :param inclusive: Include messages with latest or oldest timestamp in results
-        :type inclusive: bool
-        :param limit: The number of messages to return
-        :type limit: int
-        :param latest: End of time range to include in results
-        :type latest: str
-        :param oldest: Start of time range to include in results
-        :type oldest: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.get('conversations.history',
-                        data={'channel': channel, 'cursor': cursor,
-                              'inclusive': int(inclusive), 'limit': limit,
-                              'latest': latest, 'oldest': oldest})
-
-    def history_all(self, channel):
-        """
-        Fetches all history of messages and events from a channel
-
-        :param channel: The channel ID
-        :type channel: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        response = self.get('conversations.history', params={'channel': channel})
-        conversations = response.body.get('messages', [])
-        next_cursor = response.body.get('response_metadata', {}).get('next_cursor', '')
-        while next_cursor:
-            response = self.get('conversations.history',
-                                params={'channel':channel, 'cursor': next_cursor})
-            if response is not None:
-                conversations.extend(response.body.get('messages', []))
-                next_cursor = response.body.get('response_metadata', {}).get('next_cursor', '')
-                time.sleep(DEFAULT_API_SLEEP)
-            else:
-                raise SlackestError("Null response received. You've probably hit the rate limit.")
-
-        if conversations:
-            response.body['messages'] = conversations
-        return response
-
-    def info(self, channel, include_locale=False, include_num_members=False):
-        """
-        Gets information about a channel.
-
-        :param channel: The channel ID
-        :type channel: str
-        :param include_locale: Include the locale of the members in the channel
-        :type include_locale: bool
-        :param include_num_members: Include the number of members in the channel
-        :type include_num_members: bool
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('conversations.info', data={
-            'channel': channel,
-            'include_locale': str(include_locale).lower(),
-            'include_num_members': str(include_num_members).lower()})
-
-    def invite(self, channel, users=[]):
-        """
-        Invites users to a channel
-
-        :param name: The channel ID
-        :type name: str
-        :param users: A list of User IDs to invite to the channel
-        :type users: list[str]
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        if isinstance(users, (tuple, list)):
-            users = ','.join(users)
-        return self.post('conversations.invite', data={'channel': channel, 'users': users})
-
-    def join(self, channel):
-        """
-        Allows a user object to join a channel
-
-        :param name: The channel ID
-        :type name: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('conversations.join', data={'channel': channel})
-
-    def kick(self, channel, user):
-        """
-        Removes a user from a channel
-
-        :param channel: The channel ID
-        :type channel: str
-        :param user: The user ID
-        :type user: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('conversations.kick', data={'channel': channel, 'user': user})
-
-    def leave(self, channel):
-        """
-        Allows a user object to leave a channel
-
-        :param name: The channel ID
-        :type name: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('conversations.leave', data={'channel': channel})
-
-    def list(self, cursor=None, exclude_archived=False, limit=100, types="public_channel"):
-        """
-        Lists channels
-
-        :param cursor: the cursor id of the next set of the list
-        :type cursor: str
-        :param exclude_archived: Exclude archived channels
-        :type exclude_archived: bool
-        :param limit: The number of conversations to return
-        :type limit: int
-        :param types: The type of channel to return, can be one of public_channel, private_channel
-        :type types: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('conversations.list',
-                         data={'cursor': cursor, 'exclude_archived': str(exclude_archived).lower(),
-                               'limit': limit, 'types': types})
-
-    def list_all(self, exclude_archived=False, types="public_channel"):
-        """
-        Lists all channels
-
-        :param exclude_archived: Exclude archived channels
-        :type exclude_archived: bool
-        :param types: The type of channel to return, can be one of public_channel, private_channel
-        :type types: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        response = self.get('conversations.list',
-                            params={'exclude_archived': str(exclude_archived).lower(),
-                                    'types': types})
-        channels = response.body.get('channels', [])
-        next_cursor = response.body.get('response_metadata', {}).get('next_cursor', '')
-        while next_cursor:
-            response = self.get('conversations.list',
-                                params={'exclude_archived': str(exclude_archived).lower(),
-                                        'types': types, 'cursor': next_cursor})
-            if response is not None:
-                channels.extend(response.body.get('channels', []))
-                next_cursor = response.body.get('response_metadata', {}).get('next_cursor', '')
-                time.sleep(DEFAULT_API_SLEEP)
-            else:
-                raise SlackestError("Null response received. You've probably hit the rate limit.")
-
-        if channels:
-            response.body['channels'] = channels
-        return response
-
-    def members(self, channel, cursor=None, limit=100):
-        """
-        Lists members of a channel
-
-        :param channel: The channel ID
-        :type channel: str
-        :param cursor: the cursor id of the next set of the list
-        :type cursor: str
-        :param limit: The number of messages to return
-        :type limit: int
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('conversations.members',
-                         data={'channel': channel, 'cursor': cursor, 'limit': limit})
-
-    def members_all(self, channel):
-        """
-        Lists all members of a channel
-
-        :param channel: The channel ID
-        :type channel: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        response = self.get('conversations.members', params={'channel': channel})
-        members = response.body.get('members', [])
-        next_cursor = response.body.get('response_metadata', {}).get('next_cursor', '')
-        while next_cursor:
-            response = self.get('conversations.members',
-                                params={'channel': channel, 'cursor': next_cursor})
-            if response is not None:
-                members.extend(response.body.get('members', []))
-                next_cursor = response.body.get('response_metadata', {}).get('next_cursor', '')
-                time.sleep(DEFAULT_API_SLEEP)
-            else:
-                raise SlackestError("Null response received. You've probably hit the rate limit.")
-
-        if members:
-            response.body['members'] = members
-        return response
-
-    def open(self, channel, return_im=True, users=[]):
-        """
-        Opens or resumes DMs or multi person DMs
-
-        :param channel: The channel ID
-        :type channel: str
-        :param return_im: Indicates you wnat the full IM channel definition in the response
-        :type return_im: bool
-        :param user_ids: A list of User IDs to invite to the channel
-        :type user_ids: list[str]
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        if isinstance(users, (tuple, list)):
-            users = ','.join(users)
-
-        self.post('conversations.open',
-                  data={'channel': channel, 'return_im': str(return_im).lower(), 'users': users})
-
-    def rename(self, channel, name):
-        """
-        Renames a channel
-
-        :param channel: The channel ID to rename
-        :type channel: str
-        :param name: The new name of the channel
-        :type name: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        self.post('conversations.rename', data={'channel': channel, 'name': name})
-
-    def replies(self, channel, time_stamp, cursor=None, inclusive=False, limit=100,
-                latest=timestamp, oldest=0):
-        """
-        Fetches replies in a thread of messages
-
-        :param channel: The channel ID
-        :type channel: str
-        :param time_stamp: Unique identifier of a thread's parent message
-        :type time_stamp: str
-        :param cursor: the cursor id of the next set of replies
-        :type cursor: str
-        :param inclusive: Include messages with latest or oldest timestamp in results
-        :type inclusive: bool
-        :param limit: The number of messages to return
-        :type limit: int
-        :param latest: End of time range to include in results
-        :type latest: str
-        :param latest: Start of time range to include in results
-        :type oldest: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('conversations.replies',
-                         data={'channel': channel, 'ts': time_stamp, 'cursor': cursor,
-                               'inclusive': int(inclusive), 'limit': limit,
-                               'latest': latest, 'oldest': oldest})
-
-    def replies_all(self, channel, time_stamp):
-        """
-        Fetches all replies in a thread of messages
-
-        :param channel: The channel ID
-        :type channel: str
-        :param time_stamp: Unique identifier of a thread's parent message
-        :type time_stamp: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        response = self.get('conversations.replies',
-                            params={'channel': channel, 'ts': time_stamp})
-        replies = response.body.get('message', [])
-        next_cursor = response.body.get('response_metadata', {}).get('next_cursor', '')
-        while next_cursor:
-            response = self.get('conversations.replies',
-                                params={'channel': channel, 'ts': time_stamp,
-                                        'cursor': next_cursor})
-            if response is not None:
-                replies.extend(response.body.get('message', []))
-                next_cursor = response.body.get('response_metadata', {}).get('next_cursor', '')
-                time.sleep(DEFAULT_API_SLEEP)
-            else:
-                raise SlackestError("Null response received. You've probably hit the rate limit.")
-
-        if replies:
-            response.body['message'] = replies
-        return response
-
-    def setPurpose(self, channel, purpose):
-        """
-        Assigns purpose to a channel
-
-        :param channel: The channel ID
-        :type channel: str
-        :param purpose: The new purpose of the channel
-        :type purpose: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('conversations.setPurpose', data={'channel': channel, 'purpose': purpose})
-
-    def setTopic(self, channel, topic):
-        """
-        Assigns topic to a channel
-
-        :param channel: The channel ID
-        :type channel: str
-        :param topic: The new topic of the channel
-        :type topic: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('conversations.setTopic', data={'channel': channel, 'topic': topic})
-
-    def unarchive(self, channel):
-        """
-        Unarchives a channel
-
-        :param channel: The channel ID
-        :type channel: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('conversations.unarchive', data={'channel': channel})
-
-
-class Chat(BaseAPI):
-    """Follows the Slack Chat API. See https://api.slack.com/methods"""
-
-    def post_message(self, channel, text=None, username=None, as_user=False,
-                     parse=None, link_names=None, attachments=None,
-                     unfurl_links=None, unfurl_media=None, icon_url=None,
-                     icon_emoji=None, thread_ts=None, reply_broadcast=None):
-        """
-        Posts a message to a channel
-
-        :param channel: The channel ID
-        :type channel: str
-        :param text: Text of the message to post
-        :type text: str
-        :param username: The username to post as, must be used w/ as_user
-        :type username: str
-        :param as_user: Posts as the user instead of a bot
-        :type as_user: bool
-        :param parse: Change how messages are treated
-        :type parse: str
-        :param link_names: Find and link channel names and username
-        :type link_names: str
-        :param attachments: JSON based array of structured attachments
-        :type attachments: JSON
-        :param unfurl_links: Enable unfurling of links
-        :type unfurl_links: str
-        :param unfurl_media: Enable unfurling of media
-        :type unfurl_media: str
-        :param icon_url: The icon URL
-        :type icon_url: str
-        :param icon_emoji: Emoji to use as the icon for this message
-        :type icon_emoji: str
-        :param thread_ts: Provide another messages ts value to make this message a reply
-        :type thread_ts: str
-        :param reply_broadcast: Indicates whether reply should be visible in the channel
-        :type reply_broadcast: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        # Ensure attachments are json encoded
-        if attachments:
-            if isinstance(attachments, list):
-                attachments = json.dumps(attachments)
-
-        return self.post('chat.postMessage',
-                         data={
-                             'channel': channel,
-                             'text': text,
-                             'username': username,
-                             'as_user': str(as_user).lower(),
-                             'parse': parse,
-                             'link_names': link_names,
-                             'attachments': attachments,
-                             'unfurl_links': unfurl_links,
-                             'unfurl_media': unfurl_media,
-                             'icon_url': icon_url,
-                             'icon_emoji': icon_emoji,
-                             'thread_ts': thread_ts,
-                             'reply_broadcast': reply_broadcast
-                         })
-
-    def me_message(self, channel, text):
-        """
-        Share a me message to a channel
-
-        :param channel: The channel to post to
-        :type channel: str
-        :param text: The text of the message
-        :type text: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('chat.meMessage',
-                         data={'channel': channel, 'text': text})
-
-    def command(self, channel, command, text):
-        """
-        DEPRECATED? Run a command in a chat
-
-        :param channel: The channel ID
-        :type channel: str
-        :param command: The command to run
-        :type command: str
-        :param text: The text attached to the command
-        :type text: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('chat.command',
-                         data={
-                             'channel': channel,
-                             'command': command,
-                             'text': text
-                         })
-
-    def update(self, channel, time_stamp, text, attachments=None, parse=None,
-               link_names=None, as_user=False):
-        """
-        Updates a message in a channel
-
-        :param channel: The channel ID
-        :type channel: str
-        :param time_stamp: Timestamp of the message to be updated
-        :type time_stamp: str
-        :param text: New text for the message
-        :type text: str
-        :param attachments: JSON array of structured attachments
-        :type attachments: JSON
-        :param parse: Change hor messages are treated
-        :type parse: str
-        :param link_names: Find and link channel names
-        :type link_names: str
-        :param as_user: Update the message as the authed user
-        :type as_user: bool
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        # Ensure attachments are json encoded
-        if attachments is not None and isinstance(attachments, list):
-            attachments = json.dumps(attachments)
-        return self.post('chat.update',
-                         data={'channel': channel, 'ts': time_stamp, 'text': text,
-                               'attachments': attachments, 'parse': parse,
-                               'link_names': str(link_names).lower(),
-                               'as_user': str(as_user).lower})
-
-    def delete(self, channel, time_stamp, as_user=False):
-        """
-        Delete a message
-
-        :param channel: The channel ID
-        :type channel: str
-        :param time_stamp: Timestamp of the message to be deleted
-        :type time_stamp: str
-        :param as_user: Deletes the message as the authed user
-        :type as_user: bool
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('chat.delete',
-                         data={'channel': channel, 'ts': time_stamp,
-                               'as_user': str(as_user).lower()})
-
-    def post_ephemeral(self, channel, text, user, as_user=False,
-                       attachments=None, link_names=True, parse=None):
-        """
-        Sends an ephemeral message to a user in a channel
-
-        :param channel: The channel ID
-        :type channel: str
-        :param text: Text of the message to send
-        :type text: str
-        :param user: The user ID
-        :type user: str
-        :param as_user: Posts the message as the authed user
-        :type as_user: bool
-        :param attachments: JSON array of structured attachments
-        :type attachments: JSON
-        :param link_names: Link channel names and users
-        :type link_names: str
-        :param parse: Change how messages are treated
-        :type parse: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        # Ensure attachments are json encoded
-        if attachments is not None and isinstance(attachments, list):
-            attachments = json.dumps(attachments)
-        return self.post('chat.postEphemeral',
-                         data={
-                             'channel': channel,
-                             'text': text,
-                             'user': user,
-                             'as_user': str(as_user).lower(),
-                             'attachments': attachments,
-                             'link_names': str(link_names).lower(),
-                             'parse': str(parse).lower(),
-                         })
-
-    def unfurl(self, channel, time_stamp, unfurls, user_auth_message=None,
-               user_auth_required=False, user_auth_url=None):
-        """
-        Provides custom unfurl behavior for user posted URLS
-
-        :param channel: The channel ID
-        :type channel: str
-        :param time_stamp: Timestamp of the message to add unfurl behavior
-        :type time_stamp: str
-        :param unfurls: JSON map with keys set to URLS in the message
-        :type unfurls: JSON
-        :param user_auth_message: Invitation to user to use Slack app
-        :type user_auth_message: str
-        :param user_auth_required: Slack app required
-        :type user_auth_required: bool
-        :param user_auth_unfurl: URL for completion
-        :type user_auth_unfurl: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('chat.unfurl',
-                         data={'channel': channel, 'ts': time_stamp,
-                               'unfurls': unfurls, 'user_auth_message': user_auth_message,
-                               'user_auth_required': int(user_auth_required),
-                               'user_auth_url': user_auth_url})
-
-    def get_permalink(self, channel, message_ts):
-        """
-        Retrieve a permalink URL for a specific extant message
-
-        :param channel:
-        :type channel:
-        :param message_ts:
-        :type message_ts:
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.get('chat.getPermalink',
-                        params={'channel': channel, 'message_ts': message_ts})
-
-
-class IM(BaseAPI):
-    """Follows the Slack IM API. See https://api.slack.com/methods"""
-
-    def list(self):
-        """
-        Lists direct messages for the calling user
-
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.get('im.list')
-
-    def history(self, channel, latest=None, oldest=None, count=None,
-                inclusive=True, unreads=False):
-        """
-        Fetches history of messages and events from a DM channel
-
-        :param channel: The channel ID
-        :type channel: str
-        :param latest: End of time range of messages to include in results
-        :type latest: str
-        :param oldest: Start of time range of messages to include in results
-        :type oldest: str
-        :param count: Number of messages to return
-        :type count: int
-        :param inclusive: Include messages with oldest/latest inclusive
-        :type inclusive: bool
-        :param unreads: Include unread count display
-        :type unreads: bool
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.get('im.history',
-                        params={
-                            'channel': channel,
-                            'latest': latest,
-                            'oldest': oldest,
-                            'count': count,
-                            'inclusive': int(inclusive),
-                            'unreads' : int(unreads)
-                        })
-
-    def replies(self, channel, thread_ts):
-        """
-        Retrieves a thread of messages posted to a DM
-
-        :param channel: The channel ID
-        :type channel: str
-        :param thread_ts: Unique ID of thread's parent message
-        :type thread_ts: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.get('im.replies',
-                        params={'channel': channel, 'thread_ts': thread_ts})
-
-    def mark(self, channel, time_stamp):
-        """
-        Sets the read cursor in a DM
-
-        :param channel: The channel ID
-        :type channel: str
-        :param time_stamp: Timestamp of the most recently seen message
-        :type time_stamp: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('im.mark', data={'channel': channel, 'ts': time_stamp})
-
-    def open(self, user, include_locale=True, return_im=True):
-        """
-        Opens a DM channel
-
-        :param user:  User to open a DM channel with
-        :type user: str
-        :param include_locale: Receive locales for this DM
-        :type include_locale: str
-        :param return_im: Return the full IM channel definition
-        :type return_im: True
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('im.open',
-                         data={'user': user, 'include_locale': str(include_locale).lower(),
-                               'return_im': str(return_im).lower()})
-
-    def close(self, channel):
-        """
-        Close a DM channel
-
-        :param channel:
-        :type channel:
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('im.close', data={'channel': channel})
-
-
-class MPIM(BaseAPI):
-    """Follows the Slack MPIM API. See https://api.slack.com/methods"""
-
-    def open(self, users=[]):
-        """
-        Opens a MPIM with a list of users
-
-        :param users: A list of user IDs
-        :type users: list[str]
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        if isinstance(users, (tuple, list)):
-            users = ','.join(users)
-
-        return self.post('mpim.open', data={'users': users})
-
-    def close(self, channel):
-        """
-        Closes a MPIM
-
-        :param channel: the channel ID
-        :type channel: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('mpim.close', data={'channel': channel})
-
-    def mark(self, channel, time_stamp):
-        """
-        Sets the read cursor in a MPIM
-
-        :param channel: The channel ID
-        :type channel: str
-        :param time_stamp: The timestamp of the message
-        :type time_stamp: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('mpim.mark', data={'channel': channel, 'ts': time_stamp})
-
-    def list(self):
-        """
-        Lists MPIM for the calling user
-
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.get('mpim.list')
-
-    def history(self, channel, latest=None, oldest=None, inclusive=False,
-                count=None, unreads=False):
-        """
-        Fetches a history of messages and events
-
-        :param channel: The channel ID
-        :type channel: str
-        :param latest: End of time range to include in results
-        :type latest: str
-        :param oldest: Start of time range to include in results
-        :type oldest: str
-        :param inclusive: Include latest/oldest messags
-        :type inclusive: str
-        :param count: Number of messages to return
-        :type count: int
-        :param unreads: Include count display
-        :type unreads: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.get('mpim.history',
-                        params={
-                            'channel': channel,
-                            'latest': latest,
-                            'oldest': oldest,
-                            'inclusive': int(inclusive),
-                            'count': count,
-                            'unreads': int(unreads)
-                        })
-
-    def replies(self, channel, thread_ts):
-        """
-        Retrieves a thread of messages posted to MPIM
-
-        :param channel: The channel ID
-        :type channel: str
-        :param thread_ts: Thread's parent message
-        :type thread_ts: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.get('mpim.replies',
-                        params={'channel': channel, 'thread_ts': thread_ts})
-
-
-class Search(BaseAPI):
-    """Follows the Slack Search API. See https://api.slack.com/methods"""
-
-    def all(self, query, sort=None, sort_dir=None, highlight=True, count=None,
-            page=None):
-        """
-        Searches for messages and files matching a query
-
-        :param query: Search query
-        :type query: str
-        :param sort: Sort by score or timestamp
-        :type sort: str
-        :param sort_dir: Sort direction asc or desc
-        :type sort_dir: str
-        :param highlight: Enable highlight markers
-        :type highlight: str
-        :param count: Number of items to return
-        :type count: int
-        :param page: Page number of results to return
-        :type page: int
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.get('search.all',
-                        params={
-                            'query': query,
-                            'sort': sort,
-                            'sort_dir': sort_dir,
-                            'highlight': str(highlight).lower(),
-                            'count': count,
-                            'page': page
-                        })
-
-    def files(self, query, sort=None, sort_dir=None, highlight=True,
-              count=None, page=None):
-        """
-        Searches for files matching a query
-
-        :param query: Search query
-        :type query: str
-        :param sort: Sort by score or timestamp
-        :type sort: str
-        :param sort_dir: Sort direction asc or desc
-        :type sort_dir: str
-        :param highlight: Enable highlight markers
-        :type highlight: str
-        :param count: Number of items to return
-        :type count: int
-        :param page: Page number of results to return
-        :type page: int
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.get('search.files',
-                        params={
-                            'query': query,
-                            'sort': sort,
-                            'sort_dir': sort_dir,
-                            'highlight': str(highlight).lower(),
-                            'count': count,
-                            'page': page
-                        })
-
-    def messages(self, query, sort=None, sort_dir=None, highlight=True,
-                 count=None, page=None):
-        """
-        Searches for messages matching a query
-
-        :param query: Search query
-        :type query: str
-        :param sort: Sort by score or timestamp
-        :type sort: str
-        :param sort_dir: Sort direction asc or desc
-        :type sort_dir: str
-        :param highlight: Enable highlight markers
-        :type highlight: str
-        :param count: Number of items to return
-        :type count: int
-        :param page: Page number of results to return
-        :type page: int
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.get('search.messages',
-                        params={
-                            'query': query,
-                            'sort': sort,
-                            'sort_dir': sort_dir,
-                            'highlight': str(highlight).lower(),
-                            'count': count,
-                            'page': page
-                        })
-
-
-class FilesComments(BaseAPI):
-    """Follows the Slack FilesComments API. See https://api.slack.com/methods"""
-
-    def add(self, file_, comment):
-        """
-        DEPRECATED - Adds a comment to a file
-
-        :param file_: The file ID
-        :type file_: str
-        :param comment: Text of the comment
-        :type comment: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('files.comments.add',
-                         data={'file': file_, 'comment': comment})
-
-    def delete(self, file_, id):
-        """
-        Deletes a comment on a file
-
-        :param file_: File to delete a comment from
-        :type file_: str
-        :param id: The comment ID
-        :type id: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('files.comments.delete',
-                         data={'file': file_, 'id': id})
-
-    def edit(self, file_, id, comment):
-        """
-        DEPRECATED - Edits a comment to a file
-
-        :param file_: File to delete a comment from
-        :type file_: str
-        :param id: The comment ID
-        :type id: str
-        :param comment: Text of the comment
-        :type comment: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('files.comments.edit',
-                         data={'file': file_, 'id': id, 'comment': comment})
-
-
-class Files(BaseAPI):
-    """Follows the Slack Files API. See https://api.slack.com/methods"""
-
-    def __init__(self, *args, **kwargs):
-        super(Files, self).__init__(*args, **kwargs)
-        self._comments = FilesComments(*args, **kwargs)
-
-    @property
-    def comments(self):
-        return self._comments
-
-    def list(self, user=None, ts_from=None, ts_to=None, types=None,
-             count=None, page=None, channel=None):
-        """
-        List of files within a team
-
-        :param user: Filter files to this user ID
-        :type user: str
-        :param ts_from: Timestamp from = after
-        :type ts_from: str
-        :param ts_to: Timestamp to = before
-        :type ts_to: str
-        :param types: Filter files by type
-        :type types: str
-        :param count: Number of items to return
-        :type count: int
-        :param page: Page number of results to return
-        :type page: int
-        :param channel: Filter files to this channel ID
-        :type channel: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.get('files.list',
-                        params={
-                            'user': user,
-                            'ts_from': ts_from,
-                            'ts_to': ts_to,
-                            'types': types,
-                            'count': count,
-                            'page': page,
-                            'channel': channel
-                        })
-
-    def info(self, file_, count=None, page=None, cursor=None, limit=100):
-        """
-        Gents information about a file
-
-        :param file_: The file ID
-        :type file_: str
-        :param count: Number of items to return
-        :type count: int
-        :param page: Page number of results to return
-        :type page: int
-        :param cursor: The parameter for pagination
-        :type cursor: str
-        :param limit: Max number of items to return
-        :type limit: int
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.get('files.info',
-                        params={'file': file_, 'count': count, 'page': page,
-                                'cursor': cursor, 'limit': limit})
-
-    def upload(self, file_=None, content=None, filetype=None, filename=None,
-               title=None, initial_comment=None, channels=None, thread_ts=None):
-        """
-        Uploads or creates a file
-
-        :param file_: The file ID
-        :type file_: str
-        :param content: File contents via a POST variable
-        :type content: binary
-        :param filetype: File type identifier
-        :type filetype: str
-        :param filename: File name
-        :type filename: str
-        :param title: Title of the file
-        :type title: str
-        :param initial_comment: Comment on the file
-        :type initial_comment: str
-        :param channels: CSV of channel names to post to
-        :type channels: list[str]
-        :param thread_ts: Parent thread to use in a reply
-        :type thread_ts: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        if isinstance(channels, (tuple, list)):
-            channels = ','.join(channels)
-
-        data = {
-            'content': content,
-            'filetype': filetype,
-            'filename': filename,
-            'title': title,
-            'initial_comment': initial_comment,
-            'channels': channels,
-            'thread_ts': thread_ts
-        }
-
-        if file_:
-            if isinstance(file_, str):
-                with open(file_, 'rb') as file_name:
-                    return self.post('files.upload', data=data, files={'file': file_name})
-
-            return self.post(
-                'files.upload', data=data, files={'file': file_}
-            )
-        else:
-            return self.post('files.upload', data=data)
-
-    def delete(self, file_):
-        """
-        Deletes a file
-
-        :param file_: The file ID
-        :type file_: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('files.delete', data={'file': file_})
-
-    def revoke_public_url(self, file_):
-        """
-        Revokes public sharing
-
-        :param file_: The file ID
-        :type file_: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('files.revokePublicURL', data={'file': file_})
-
-    def shared_public_url(self, file_):
-        """
-        Enables public sharing
-
-        :param file_: The file ID
-        :type file_: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.post('files.sharedPublicURL', data={'file': file_})
-
-
-class Stars(BaseAPI):
-    """Follows the Slack Stars API. See https://api.slack.com/methods"""
-
-    def add(self, file_=None, file_comment=None, channel=None, timestamp=None):
-        """
-        Adds a star to an item
-
-        :param file_: The file ID
-        :type file_: str
-        :param file_comment: The comment on the file
-        :type file_comment: str
-        :param channel: The channel ID
-        :type channel: str
-        :param timestamp: The timestamp of the message
-        :type timestamp: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        assert file_ or file_comment or channel
-
-        return self.post('stars.add',
-                         data={
-                             'file': file_,
-                             'file_comment': file_comment,
-                             'channel': channel,
-                             'timestamp': timestamp
-                         })
-
-    def list(self, user=None, count=None, page=None):
-        """
-        Lists stars for a user
-
-        :param :
-        :type :
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.get('stars.list',
-                        params={'user': user, 'count': count, 'page': page})
-
-    def remove(self, file_=None, file_comment=None, channel=None, timestamp=None):
-        """
-        Removes a star from an item
-
-        :param file_: The file ID
-        :type file_: str
-        :param file_comment: The comment on the file
-        :type file_comment: str
-        :param channel: The channel ID
-        :type channel: str
-        :param timestamp: The timestamp of the message
-        :type timestamp: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        assert file_ or file_comment or channel
-
-        return self.post('stars.remove',
-                         data={
-                             'file': file_,
-                             'file_comment': file_comment,
-                             'channel': channel,
-                             'timestamp': timestamp
-                         })
-
-
-class Emoji(BaseAPI):
-    """Follows the Slack Emoji API. See https://api.slack.com/methods"""
-
-    def list(self):
-        """
-        List all emojis
-
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.get('emoji.list')
-
-
-class Presence(BaseAPI):
-    """Follows the Slack Presence API. See https://api.slack.com/methods"""
-
-    AWAY = 'away'
-    ACTIVE = 'active'
-    TYPES = (AWAY, ACTIVE)
-
-    def set(self, presence):
-        """
-        DEPRECATED - Sets the precense of a user
-
-        :param presence: The status
-        :type presence: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        assert presence in Presence.TYPES, 'Invalid presence type'
-        return self.post('presence.set', data={'presence': presence})
-
-
-class RTM(BaseAPI):
-    """Follows the Slack RTM API. See https://api.slack.com/methods"""
-
-    def start(self, simple_latest=True, no_unreads=False, mpim_aware=False):
-        """
-        Start a Real Time Messaging session
-
-        :param simple_latest: Return timestamp only for latest message object
-        :type simple_latest: bool
-        :param no_unreads: Skip unread counts
-        :type no_unreads: bool
-        :param mpim_aware: Returns MPIMs to the client
-        :type mpim_aware: bool
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.get('rtm.start',
-                        params={
-                            'simple_latest': str(simple_latest).lower(),
-                            'no_unreads': str(no_unreads).lower(),
-                            'mpim_aware': str(mpim_aware).lower(),
-                        })
-
-    def connect(self):
-        """
-        Start a Real Time Messaging session
-
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return self.get('rtm.connect')
-
-
-class TeamProfile(BaseAPI):
-    """Follows the Slack TeamProfile API. See https://api.slack.com/methods"""
-
-    def get(self, visibility=None):
-        """
-        Retrieves a team's profile
-
-        :param visibility: Filter by visibility
-        :type visibility: str
-        :return: A response object to run the API request.
-        :rtype: :class:`Response <Response>` object
-        """
-        return super(TeamProfile, self).get(
-            'team.profile.get',
-            params={'visibility': str(visibility).lower()}
-        )
-
+from .team_profile import TeamProfile
 
 class Team(BaseAPI):
     """Follows the Slack Team API. See https://api.slack.com/methods"""
@@ -2853,29 +760,7 @@ class Apps(BaseAPI):
         return self._permissions
 
 
-class IncomingWebhook(object):
-    """Follows the Slack IncomingWebhook API. See https://api.slack.com/methods"""
-
-    def __init__(self, url=None, timeout=DEFAULT_TIMEOUT, proxies=None):
-        self.url = url
-        self.timeout = timeout
-        self.proxies = proxies
-
-    def post(self, data):
-        """
-        Posts message with payload formatted in accordance with
-        this documentation https://api.slack.com/incoming-webhooks
-
-        :param data: The data payload
-        :type data: A JSON representation of the payload
-        :return: A response object to run the request.
-        :rtype: :class:`Response <Response>` object
-        """
-        if not self.url:
-            raise SlackestError('URL for incoming webhook is undefined')
-
-        return requests.post(self.url, data=json.dumps(data),
-                             timeout=self.timeout, proxies=self.proxies)
+from .incoming_webhook import IncomingWebhook
 
 
 class Slackest(object):
