+++ slackest/constants.py
@@ -1,0 +1,6 @@
+API_BASE_URL = 'https://slack.com/api/{api}'
+DEFAULT_TIMEOUT = 10
+DEFAULT_RETRIES = 0
+# seconds to wait after a 429 error if Slack's API doesn't provide one
+DEFAULT_WAIT = 20
+DEFAULT_API_SLEEP = 5
