+++ slackest/slackest_error.py
@@ -1,0 +1,3 @@
+class SlackestError(Exception):
+    """Dummy exception placeholder"""
+    pass
