+++ tests/test_slackest.py
@@ -11,7 +11,7 @@ from slackest import Slackest
 
 class TestUtils(unittest.TestCase):
     # create_channel
-    @patch('slackest.requests')
+    @patch('slackest.base_api.requests')
     def test_create_channel(self, mock_requests):
         json_to_text = json.dumps(stubs.good_channel_info_response)
         mock_requests.post.return_value = Mock(
@@ -21,7 +21,7 @@ class TestUtils(unittest.TestCase):
         new_channel_id = sc.create_channel('busting', False, ["U0G9QF9C6", "U1QNSQB9U"]).body['channel'].get('id')
         self.assertEqual('C1H9RESGL', new_channel_id)
 
-    @patch('slackest.requests')
+    @patch('slackest.base_api.requests')
     def test_create_channel_error(self, mock_requests):
         json_to_text = json.dumps(stubs.bad_channel_info_response)
         mock_requests.post.return_value = Mock(
@@ -32,7 +32,7 @@ class TestUtils(unittest.TestCase):
         self.assertEqual('channel_not_found', error)
 
     # get_channels
-    @patch('slackest.requests')
+    @patch('slackest.base_api.requests')
     def test_get_channels(self, mock_requests):
         json_to_text = json.dumps(stubs.good_conversation_list_response)
         mock_requests.get.return_value = Mock(
@@ -42,7 +42,7 @@ class TestUtils(unittest.TestCase):
         channel_id = sc.get_channels(False, ['public_channel','private_channel','mpim','im']).body['channels'][0].get('id')
         self.assertEqual('G0AKFJBEU', channel_id)
 
-    @patch('slackest.requests')
+    @patch('slackest.base_api.requests')
     def test_get_channels_error(self, mock_requests):
         json_to_text = json.dumps(stubs.bad_conversation_list_response)
         mock_requests.get.return_value = Mock(
@@ -53,7 +53,7 @@ class TestUtils(unittest.TestCase):
         self.assertEqual('invalid_auth', error)
 
     # list_all_users
-    @patch('slackest.requests')
+    @patch('slackest.base_api.requests')
     def test_list_all_users(self, mock_requests):
         json_to_text = json.dumps(stubs.good_user_list_response)
         mock_requests.get.return_value = Mock(
@@ -64,7 +64,7 @@ class TestUtils(unittest.TestCase):
         first_user_id = all_users[0].get('id')
         self.assertEqual('W012A3CDE', first_user_id)
 
-    @patch('slackest.requests')
+    @patch('slackest.base_api.requests')
     def test_list_all_users_error(self, mock_requests):
         json_to_text = json.dumps(stubs.bad_user_list_response)
         mock_requests.get.return_value = Mock(
@@ -75,7 +75,7 @@ class TestUtils(unittest.TestCase):
         self.assertEqual('fatal_error', error)
 
     # kick_user
-    @patch('slackest.requests')
+    @patch('slackest.base_api.requests')
     def test_kick_user(self, mock_requests):
         json_to_text = json.dumps(stubs.good_kick_user_response)
         mock_requests.post.return_value = Mock(
@@ -85,7 +85,7 @@ class TestUtils(unittest.TestCase):
         ku = sc.kick_user('C0G9QF9GW', 'U0G9QF9C6').body['ok']
         self.assertEqual('true', ku)
 
-    @patch('slackest.requests')
+    @patch('slackest.base_api.requests')
     def test_kick_user_error(self, mock_requests):
         json_to_text = json.dumps(stubs.bad_kick_user_response)
         mock_requests.post.return_value = Mock(
@@ -96,7 +96,7 @@ class TestUtils(unittest.TestCase):
         self.assertEqual('user_not_found', error)
 
     # history_all
-    @patch('slackest.requests')
+    @patch('slackest.base_api.requests')
     def test_get_conversation_history(self, mock_requests):
         json_to_text = json.dumps(stubs.good_conversation_history_response)
         mock_requests.get.return_value = Mock(
@@ -106,7 +106,7 @@ class TestUtils(unittest.TestCase):
         second_user = sc.history_all('asd').body['messages'][1].get('user')
         self.assertEqual('U061F7AUR', second_user)
 
-    @patch('slackest.requests')
+    @patch('slackest.base_api.requests')
     def test_get_conversation_history_error(self, mock_requests):
         json_to_text = json.dumps(stubs.bad_conversation_history_response)
         mock_requests.get.return_value = Mock(
@@ -117,7 +117,7 @@ class TestUtils(unittest.TestCase):
         self.assertEqual('channel_not_found', error)
 
     # post_message_to_channel
-    @patch('slackest.requests')
+    @patch('slackest.base_api.requests')
     def test_post_message_to_channel(self, mock_requests):
         json_to_text = json.dumps(stubs.good_conversation_history_response)
         mock_requests.post.return_value = Mock(
@@ -127,7 +127,7 @@ class TestUtils(unittest.TestCase):
         message_user = sc.post_message_to_channel('slimer', "Here's a message for you").body['messages'][0].get('user')
         self.assertEqual('U012AB3CDE', message_user)
 
-    @patch('slackest.requests')
+    @patch('slackest.base_api.requests')
     def test_post_message_to_channel_error(self, mock_requests):
         json_to_text = json.dumps(stubs.bad_conversation_history_response)
         mock_requests.post.return_value = Mock(
@@ -138,7 +138,7 @@ class TestUtils(unittest.TestCase):
         self.assertEqual('channel_not_found', error)
 
     # post_thread_to_message
-    @patch('slackest.requests')
+    @patch('slackest.base_api.requests')
     def test_post_thread_to_message(self, mock_requests):
         json_to_text = json.dumps(stubs.good_post_thread_to_message_response)
         mock_requests.post.return_value = Mock(
@@ -148,7 +148,7 @@ class TestUtils(unittest.TestCase):
         username = sc.post_thread_to_message('slimer', "Here's a message for you", '1503435956.000247').body['message'].get('username')
         self.assertEqual('ecto1', username)
 
-    @patch('slackest.requests')
+    @patch('slackest.base_api.requests')
     def test_post_thread_to_message_error(self, mock_requests):
         json_to_text = json.dumps(stubs.bad_post_thread_to_message_response)
         mock_requests.post.return_value = Mock(
@@ -159,7 +159,7 @@ class TestUtils(unittest.TestCase):
         self.assertEqual('is_archived', error)
 
     # add_member_to_channel
-    @patch('slackest.requests')
+    @patch('slackest.base_api.requests')
     def test_add_member_to_channel(self, mock_requests):
         json_to_text = json.dumps(stubs.good_add_member_to_channel_response)
         mock_requests.post.return_value = Mock(
@@ -169,7 +169,7 @@ class TestUtils(unittest.TestCase):
         name = sc.add_member_to_channel('C012AB3CD', 'U2345678901').body['channel'].get('name')
         self.assertEqual('general', name)
 
-    @patch('slackest.requests')
+    @patch('slackest.base_api.requests')
     def test_add_member_to_channel_error(self, mock_requests):
         json_to_text = json.dumps(stubs.bad_add_member_to_channel_response)
         mock_requests.post.return_value = Mock(
@@ -180,7 +180,7 @@ class TestUtils(unittest.TestCase):
         self.assertEqual('method_not_supported_for_channel_type', error)
 
     # get_channel_info
-    @patch('slackest.requests')
+    @patch('slackest.base_api.requests')
     def test_get_channel_info(self, mock_requests):
         json_to_text = json.dumps(stubs.good_channel_info_response)
         mock_requests.get.return_value = Mock(
@@ -190,7 +190,7 @@ class TestUtils(unittest.TestCase):
         channel_name = sc.get_channel_info('C0G9QF9GW').body['channel'].get('name')
         self.assertEqual('busting', channel_name)
 
-    @patch('slackest.requests')
+    @patch('slackest.base_api.requests')
     def test_get_channel_info_error(self, mock_requests):
         json_to_text = json.dumps(stubs.bad_channel_info_response)
         mock_requests.get.return_value = Mock(
@@ -201,7 +201,7 @@ class TestUtils(unittest.TestCase):
         self.assertEqual('channel_not_found', error)
 
     # get_replies
-    @patch('slackest.requests')
+    @patch('slackest.base_api.requests')
     def test_get_replies(self, mock_requests):
         json_to_text = json.dumps(stubs.good_get_replies_response)
         mock_requests.get.return_value = Mock(
@@ -211,7 +211,7 @@ class TestUtils(unittest.TestCase):
         replies = sc.get_replies('C0G9QF9GW', '1482960137.003543').body['messages'][0].get['reply_count']
         self.assertEqual(3, replies)
 
-    @patch('slackest.requests')
+    @patch('slackest.base_api.requests')
     def test_get_replies(self, mock_requests):
         json_to_text = json.dumps(stubs.bad_get_replies_response)
         mock_requests.get.return_value = Mock(
@@ -222,7 +222,7 @@ class TestUtils(unittest.TestCase):
         self.assertEqual('thread_not_found', error)
 
     # set_purpose
-    @patch('slackest.requests')
+    @patch('slackest.base_api.requests')
     def test_set_purpose(self, mock_requests):
         json_to_text = json.dumps(stubs.good_set_purpose_response)
         mock_requests.post.return_value = Mock(
@@ -232,7 +232,7 @@ class TestUtils(unittest.TestCase):
         purpose = sc.set_purpose('C0G9QF9GW', 'My purpose').body['purpose']
         self.assertEqual('My purpose', purpose)
 
-    @patch('slackest.requests')
+    @patch('slackest.base_api.requests')
     def test_set_purpose_error(self, mock_requests):
         json_to_text = json.dumps(stubs.bad_set_purpose_response)
         mock_requests.post.return_value = Mock(
@@ -243,7 +243,7 @@ class TestUtils(unittest.TestCase):
         self.assertEqual('too_long', error)
 
     # set_topic
-    @patch('slackest.requests')
+    @patch('slackest.base_api.requests')
     def test_set_topic(self, mock_requests):
         json_to_text = json.dumps(stubs.good_set_topic_response)
         mock_requests.post.return_value = Mock(
@@ -253,7 +253,7 @@ class TestUtils(unittest.TestCase):
         topic = sc.set_topic('C0G9QF9GW', 'My topic').body['topic']
         self.assertEqual('My topic', topic)
 
-    @patch('slackest.requests')
+    @patch('slackest.base_api.requests')
     def test_set_topic_error(self, mock_requests):
         json_to_text = json.dumps(stubs.bad_set_topic_response)
         mock_requests.post.return_value = Mock(
@@ -264,7 +264,7 @@ class TestUtils(unittest.TestCase):
         self.assertEqual('user_is_restricted', error)
 
     # upload_file
-    @patch('slackest.requests')
+    @patch('slackest.base_api.requests')
     def test_upload_file(self, mock_requests):
         json_to_text = json.dumps(stubs.good_upload_file_response)
         mock_requests.post.return_value = Mock(
@@ -274,7 +274,7 @@ class TestUtils(unittest.TestCase):
         upload = sc.upload_file('README.rst', 'C0G9QF9GW').body['file'].get('id')
         self.assertEqual('F0TD00400', upload)
 
-    @patch('slackest.requests')
+    @patch('slackest.base_api.requests')
     def test_upload_file_error(self, mock_requests):
         json_to_text = json.dumps(stubs.bad_upload_file_response)
         mock_requests.post.return_value = Mock(
